requirements.txt
apache-airflow==3.0.2
apache-airflow-providers-postgres==5.11.0
psycopg2-binary